Adds config/lib/graphql
