+++ test/unit/lib/graphql.test.js
@@ -7,13 +7,17 @@ const sinon = require('sinon');
 const sinonChai = require('sinon-chai');
 
 const graphqlRequest = require('graphql-request');
+const stubs = require('../../helpers/stubs');
 const topicFactory = require('../../helpers/factories/topic');
+const config = require('../../../config/lib/graphql');
 
 chai.should();
 chai.use(sinonChai);
 
 // module to be tested
 const graphql = require('../../../lib/graphql');
+
+const fetchTopicbyIdQuery = config.queries.fetchTopicById;
 
 const sandbox = sinon.sandbox.create();
 
@@ -22,13 +26,26 @@ test.afterEach(() => {
 });
 
 // fetchTopicById
-test('fetchTopicById should execute GraphQL request and return result.topic', async () => {
+test('fetchTopicById should call request with query config and id variable and return response.topic', async () => {
   const topic = topicFactory.getValidTopic();
-  const topicId = topic.id;
-  sandbox.stub(graphqlRequest, 'request')
+  const id = topic.id;
+  sandbox.stub(graphql, 'request')
     .returns(Promise.resolve({ topic }));
 
-  const result = await graphql.fetchTopicById(topicId);
-  graphqlRequest.request.should.have.been.called;
+  const result = await graphql.fetchTopicById(id);
+  graphql.request.should.have.been.calledWith(fetchTopicbyIdQuery, { id });
   result.should.deep.equal(topic);
+});
+
+// fetchTopicById
+test('request should call graphql-request.request with given query and variables', async () => {
+  const variables = { id: '123' };
+  const response = { name: stubs.getRandomWord() };
+  sandbox.stub(graphqlRequest, 'request')
+    .returns(Promise.resolve(response));
+
+  const result = await graphql.request(fetchTopicbyIdQuery, variables);
+  graphqlRequest.request
+    .should.have.been.calledWith(`${config.url}/graphql`, fetchTopicbyIdQuery, variables);
+  result.should.deep.equal(response);
 });
